stock fundamentals category completed

+++ P3.FundamentalData.API/Models/InternationalFilings.cs
@@ -20,29 +20,29 @@ namespace P3.FundamentalData.API.Models
         public string Cik { get; set; }
         public DateTime? FillingDate { get; set; }
         public DateTime? AcceptedDate { get; set; }
-        public decimal? Revenue { get; set; }
-        public decimal? CostOfRevenue { get; set; }
-        public decimal? GrossProfit { get; set; }
+        public double? Revenue { get; set; }
+        public double? CostOfRevenue { get; set; }
+        public double? GrossProfit { get; set; }
         public double? GrossProfitRatio { get; set; }
-        public decimal? ResearchAndDevelopmentExpenses { get; set; }
-        public decimal? GeneralAndAdministrativeExpenses { get; set; }
-        public decimal? SellingAndMarketingExpenses { get; set; }
-        public decimal? SellingGeneralAndAdministrativeExpenses { get; set; }
-        public decimal? OtherExpenses { get; set; }
-        public decimal? OperatingExpenses { get; set; }
-        public decimal? CostAndExpenses { get; set; }
-        public decimal? InterestIncome { get; set; }
-        public decimal? InterestExpense { get; set; }
+        public double? ResearchAndDevelopmentExpenses { get; set; }
+        public double? GeneralAndAdministrativeExpenses { get; set; }
+        public double? SellingAndMarketingExpenses { get; set; }
+        public double? SellingGeneralAndAdministrativeExpenses { get; set; }
+        public double? OtherExpenses { get; set; }
+        public double? OperatingExpenses { get; set; }
+        public double? CostAndExpenses { get; set; }
+        public double? InterestIncome { get; set; }
+        public double? InterestExpense { get; set; }
         public long? DepreciationAndAmortization { get; set; }
         public long? Ebitda { get; set; }
         public double? Ebitdaratio { get; set; }
-        public decimal? OperatingIncome { get; set; }
+        public double? OperatingIncome { get; set; }
         public double? OperatingIncomeRatio { get; set; }
-        public decimal? TotalOtherIncomeExpensesNet { get; set; }
-        public decimal? IncomeBeforeTax { get; set; }
+        public double? TotalOtherIncomeExpensesNet { get; set; }
+        public double? IncomeBeforeTax { get; set; }
         public double? IncomeBeforeTaxRatio { get; set; }
-        public decimal? IncomeTaxExpense { get; set; }
-        public decimal? NetIncome { get; set; }
+        public double? IncomeTaxExpense { get; set; }
+        public double? NetIncome { get; set; }
         public double? NetIncomeRatio { get; set; }
         public float? Eps { get; set; }
         public float? Epsdiluted { get; set; }
